fix(openshell): gateway not initialized when extensions are started
Extract OpenshellCliBase abstract class with shared CLI plumbing and
move gateway commands (add/remove/select/list, endpoint status, gateway
status) into OpenshellGatewayCli. OpenshellCli now extends the base and
delegates gateway queries to OpenshellGatewayCli. OpenshellGateway
depends on OpenshellGatewayCli instead of the full OpenshellCli, and
gates CLI access behind checkAvailable().

Fixes #2371
